MB-31113 Additional changes for timer context capping
Additional moving uncaught exception related logging to debug level to
avoid spam in eventing.log

Reviewed-on: http://review.couchbase.org/99180
Tested-by: Abhishek Singh <abhishek@couchbase.com>

+++ common/common.go
@@ -275,7 +275,7 @@ type HandlerConfig struct {
 	SourceBucket             string
 	StatsLogInterval         int
 	StreamBoundary           DcpStreamBoundary
-	TimerContextSize         int
+	TimerContextSize         int64
 	TimerStorageRoutineCount int
 	TimerStorageChanSize     int
 	TimerQueueMemCap         int64
