+++ v8_consumer/include/client.h
@@ -26,7 +26,7 @@ const int HEADER_FRAGMENT_SIZE = 4;  // uint32
 const int PAYLOAD_FRAGMENT_SIZE = 4; // uint32
 const int SIZEOF_UINT32 = 4;
 
-int timer_context_size;
+int64_t timer_context_size;
 
 typedef struct resp_msg_s {
   std::string msg;
